Consolidate UpgradeStores implemenation into one in database
scabbard and receipt store upgrades both implemented slightly
different UpgradeStores structs. This commit combines them into
one implementation. This will also be used for future state
migrate commands.
