+++ cli/src/action/database/upgrade/receipt_store.rs
@@ -17,16 +17,10 @@
 use std::fmt::Write;
 use std::path::Path;
 
-use diesel::r2d2::{ConnectionManager, Pool};
 use openssl::hash::{hash, MessageDigest};
-use sawtooth::receipt::store::{diesel::DieselReceiptStore, lmdb::LmdbReceiptStore, ReceiptStore};
-use splinter::{
-    admin::store::{diesel::DieselAdminServiceStore, AdminServiceStore},
-    node_id::store::{diesel::DieselNodeIdStore, NodeIdStore},
-};
+use sawtooth::receipt::store::{lmdb::LmdbReceiptStore, ReceiptStore};
 
-use super::ConnectionUri;
-
+use crate::action::database::{stores::new_upgrade_stores, ConnectionUri};
 use crate::error::CliError;
 
 /// Migrate all of the transaction receipts to the `ReceiptStore`.
@@ -34,7 +28,8 @@ pub(super) fn upgrade_scabbard_receipt_store(
     receipt_db_dir: &Path,
     database_uri: &ConnectionUri,
 ) -> Result<(), CliError> {
-    let upgrade_stores = new_upgrade_stores(database_uri)?;
+    let upgrade_stores =
+        new_upgrade_stores(database_uri).map_err(|e| CliError::ActionError(format!("{}", e)))?;
 
     let node_id = if let Some(node_id) = upgrade_stores
         .new_node_id_store()
@@ -140,97 +135,4 @@ fn to_hex(bytes: &[u8]) -> String {
     }
 
     buf
-}
-
-trait UpgradeStores {
-    fn new_admin_service_store(&self) -> Box<dyn AdminServiceStore>;
-
-    fn new_node_id_store(&self) -> Box<dyn NodeIdStore>;
-
-    fn new_receipt_store(&self, circuit_id: &str, service_id: &str) -> Box<dyn ReceiptStore>;
-}
-
-fn new_upgrade_stores(database_uri: &ConnectionUri) -> Result<Box<dyn UpgradeStores>, CliError> {
-    match database_uri {
-        #[cfg(feature = "postgres")]
-        ConnectionUri::Postgres(url) => {
-            let connection_manager = ConnectionManager::<diesel::pg::PgConnection>::new(url);
-            let pool = Pool::builder().build(connection_manager).map_err(|err| {
-                CliError::ActionError(format!("Failed to build connection pool: {}", err))
-            })?;
-            // Test the connection
-            let _conn = pool
-                .get()
-                .map_err(|e| CliError::ActionError(format!("{}", e)))?;
-            Ok(Box::new(PostgresUpgradeStores(pool)))
-        }
-        #[cfg(feature = "sqlite")]
-        ConnectionUri::Sqlite(conn_str) => {
-            if (conn_str != ":memory:") && !std::path::Path::new(&conn_str).exists() {
-                return Err(CliError::ActionError(format!(
-                    "Database file '{}' does not exist",
-                    conn_str
-                )));
-            }
-            let connection_manager =
-                ConnectionManager::<diesel::sqlite::SqliteConnection>::new(conn_str);
-            let mut pool_builder = Pool::builder();
-            // A new database is created for each connection to the in-memory SQLite
-            // implementation; to ensure that the resulting stores will operate on the same
-            // database, only one connection is allowed.
-            if conn_str == ":memory:" {
-                pool_builder = pool_builder.max_size(1);
-            }
-            let pool = pool_builder.build(connection_manager).map_err(|err| {
-                CliError::ActionError(format!("Failed to build connection pool: {}", err))
-            })?;
-            // Test the connection
-            let _conn = pool
-                .get()
-                .map_err(|e| CliError::ActionError(format!("{}", e)))?;
-            Ok(Box::new(SqliteUpgradeStores(pool)))
-        }
-    }
-}
-
-#[cfg(feature = "postgres")]
-struct PostgresUpgradeStores(Pool<ConnectionManager<diesel::pg::PgConnection>>);
-
-#[cfg(feature = "postgres")]
-impl UpgradeStores for PostgresUpgradeStores {
-    fn new_admin_service_store(&self) -> Box<dyn AdminServiceStore> {
-        Box::new(DieselAdminServiceStore::new(self.0.clone()))
-    }
-
-    fn new_node_id_store(&self) -> Box<dyn NodeIdStore> {
-        Box::new(DieselNodeIdStore::new(self.0.clone()))
-    }
-
-    fn new_receipt_store(&self, circuit_id: &str, service_id: &str) -> Box<dyn ReceiptStore> {
-        Box::new(DieselReceiptStore::new(
-            self.0.clone(),
-            Some(format!("{}::{}", circuit_id, service_id)),
-        ))
-    }
-}
-
-#[cfg(feature = "sqlite")]
-struct SqliteUpgradeStores(Pool<ConnectionManager<diesel::SqliteConnection>>);
-
-#[cfg(feature = "sqlite")]
-impl UpgradeStores for SqliteUpgradeStores {
-    fn new_admin_service_store(&self) -> Box<dyn AdminServiceStore> {
-        Box::new(DieselAdminServiceStore::new(self.0.clone()))
-    }
-
-    fn new_node_id_store(&self) -> Box<dyn NodeIdStore> {
-        Box::new(DieselNodeIdStore::new(self.0.clone()))
-    }
-
-    fn new_receipt_store(&self, circuit_id: &str, service_id: &str) -> Box<dyn ReceiptStore> {
-        Box::new(DieselReceiptStore::new(
-            self.0.clone(),
-            Some(format!("{}::{}", circuit_id, service_id)),
-        ))
-    }
 }
